+++ anybox/recipe/openerp/vcs/base.py
@@ -40,6 +40,32 @@ clone_check_output = wrap_check_call(CloneError, utils.check_output)
 
 
 class BaseRepo(object):
+    """The common interface that all repository classes implement.
+
+    :param target_dir: the local directory which will serve as a working tree
+    :param offline: if ``True``, the repository instance will perform no
+                    network operation, and will fail instead if a non
+                    available revision is required.
+    :param clear_locks: Some VCS systems can leave locks after some failures
+                        and provide a separate way to break them. If ``True``,
+                        the repo
+                        will break any locks prior to operations (mostly useful
+                        for automated agents, such as CI robots)
+    :param clear_retry: if ``True`` failed updates by calling the instance are
+                        cleared (see :meth:`clear_target`) and retried once.
+                        This is intended for brittle VCSes from CI robots.
+
+    Other options depend on the concrete repository class.
+
+    Repository instances are **callable**. For each of them::
+
+        repo(rev)
+
+    will take all the steps necessary so that its local directory is a clone of
+    the remote source, at the specified revision. If needed and possible
+    The revision format depends on the
+    concrete class, but it is passed as a :class:`str`.
+    """
 
     def __init__(self, target_dir, url, clear_retry=False,
                  offline=False, clear_locks=False, **options):
@@ -54,18 +80,23 @@ class BaseRepo(object):
         self.options = options
 
     def clear_target(self):
+        """Entirely remove the target directory."""
         shutil.rmtree(self.target_dir)
 
     def clean(self):
         """Remove unwanted untracked files.
 
-        The default implementation removes Python object files and
+        This default implementation removes Python object files and
         (resulting) empty directories.
         Subclasses are supposed to implement better vcs-specific behaviours.
         It is important for release-related options that this cleaning does not
         appear as a local modification.
         """
         utils.clean_object_files(self.target_dir)
+
+    def revert(self, revision):
+        """Revert any local changes, including pending merges."""
+        raise NotImplementedError
 
     def __call__(self, revision):
         """Create if needed from remote source, and put it at wanted revision.
@@ -81,6 +112,24 @@ class BaseRepo(object):
             self.clear_target()
             self.get_update(revision)
         return self  # nicer in particular for tests
+
+    def get_update(self, revision):
+        """Make it so that the target directory is at the prescribed revision.
+
+        The target directory need not to be initialized: this method will
+        "clone" it from the remote source (whatever that means in the
+        considered VCS).
+
+        This method can fail under various circumstances, for instance if the
+        wanted revision does not exist locally and offline mode has been
+        selected.
+
+        :raises CloneError: if initial cloning fails
+        :raises UpdateError: if update of existing repository fails
+
+        Must be implemented in concrete subclasses
+        """
+        raise NotImplementedError
 
     @classmethod
     def is_versioned(cls, path):
@@ -109,12 +158,17 @@ class BaseRepo(object):
         return target_dir
 
     def uncommitted_changes(self):
-        """True if we have uncommitted changes."""
+        """True if we have uncommitted changes.
+
+        Must be implemented by concrete subclasses
+        """
         raise NotImplementedError
 
     def parents(self):
         """Return universal identifier for parent nodes, aka current revisions.
 
         There might be more than one with some VCSes (ex: pending merge in hg).
+
+        Must be implemented by concrete subclasses
         """
         raise NotImplementedError