+++ anybox/recipe/openerp/vcs/tests/test_bzr.py
@@ -180,6 +180,7 @@ class BzrTestCase(BzrBaseTestCase):
             os.chdir(self.src_repo)
             os.mkdir('subdir')
             subprocess.check_call(['bzr', 'add'])
+            subprocess.call(['bzr', 'whoami', '--branch', COMMIT_USER_FULL])
             subprocess.check_call(['bzr', 'commit', '-m', "bidule"])
         target_dir = os.path.join(self.dst_dir, "clone to clean")
         branch = BzrBranch(target_dir, self.src_repo)
@@ -221,6 +222,21 @@ class BzrTestCase(BzrBaseTestCase):
         with open(os.path.join(target_dir, 'unknownfile'), 'w') as f:
             f.write('some change')
         self.assertTrue(branch.uncommitted_changes())
+
+    def test_revert(self):
+        target_dir = os.path.join(self.dst_dir, "clone to clean")
+        branch = BzrBranch(target_dir, self.src_repo)
+        branch('last:1')
+
+        path = os.path.join(target_dir, 'tracked')
+        with open(path, 'r') as f:
+            original = f.readlines()
+        with open(path, 'w') as f:
+            f.write('a local mod')
+
+        branch.revert('last:1')
+        with open(path, 'r') as f:
+            self.assertEqual(f.readlines(), original)
 
     def test_archive(self):
         target_dir = os.path.join(self.dst_dir, "clone to archive")
@@ -339,6 +355,7 @@ class BzrTestCase(BzrBaseTestCase):
         f.write("content" + os.linesep)
         f.close()
         subprocess.call(['bzr', 'add'])
+        subprocess.call(['bzr', 'whoami', '--branch', COMMIT_USER_FULL])
         subprocess.call(['bzr', 'commit', '-m', 'poposal commit'])
         target_dir = os.path.join(self.dst_dir, "branch with merge")
         BzrBranch(target_dir, self.src_repo)('last:1')